+++ src/libraries/System.Runtime.InteropServices.JavaScript/src/System/Runtime/InteropServices/JavaScript/JSHostImplementation.cs
@@ -13,20 +13,11 @@ namespace System.Runtime.InteropServices.JavaScript
     internal static partial class JSHostImplementation
     {
         private const string TaskGetResultName = "get_Result";
-        private static readonly MethodInfo s_taskGetResultMethodInfo = typeof(Task<>).GetMethod(TaskGetResultName)!;
+        private static MethodInfo? s_taskGetResultMethodInfo;
         // we use this to maintain identity of JSHandle for a JSObject proxy
         public static readonly Dictionary<int, WeakReference<JSObject>> s_csOwnedObjects = new Dictionary<int, WeakReference<JSObject>>();
         // we use this to maintain identity of GCHandle for a managed object
         public static Dictionary<object, IntPtr> s_gcHandleFromJSOwnedObject = new Dictionary<object, IntPtr>(ReferenceEqualityComparer.Instance);
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static void RegisterCSOwnedObject(JSObject proxy)
-        {
-            lock (s_csOwnedObjects)
-            {
-                s_csOwnedObjects[(int)proxy.JSHandle] = new WeakReference<JSObject>(proxy, trackResurrection: true);
-            }
-        }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static void ReleaseCSOwnedObject(nint jsHandle)
@@ -41,8 +32,7 @@ namespace System.Runtime.InteropServices.JavaScript
             }
         }
 
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static object? GetTaskResult(Task task)
+        public static object? GetTaskResultDynamic(Task task)
         {
             MethodInfo method = GetTaskResultMethodInfo(task.GetType());
             if (method != null)
@@ -91,145 +81,6 @@ namespace System.Runtime.InteropServices.JavaScript
             return temp.methodHandle;
         }
 
-        public static MarshalType GetMarshalTypeFromType(Type type)
-        {
-            if (type is null)
-                return MarshalType.VOID;
-
-            var typeCode = Type.GetTypeCode(type);
-            if (type.IsEnum)
-            {
-                switch (typeCode)
-                {
-                    case TypeCode.Int32:
-                    case TypeCode.UInt32:
-                        return MarshalType.ENUM;
-                    case TypeCode.Int64:
-                    case TypeCode.UInt64:
-                        return MarshalType.ENUM64;
-                    default:
-                        throw new JSException($"Unsupported enum underlying type {typeCode}");
-                }
-            }
-
-            switch (typeCode)
-            {
-                case TypeCode.SByte:
-                case TypeCode.Int16:
-                case TypeCode.Int32:
-                    return MarshalType.INT;
-                case TypeCode.Byte:
-                case TypeCode.UInt16:
-                case TypeCode.UInt32:
-                    return MarshalType.UINT32;
-                case TypeCode.Boolean:
-                    return MarshalType.BOOL;
-                case TypeCode.Int64:
-                    return MarshalType.INT64;
-                case TypeCode.UInt64:
-                    return MarshalType.UINT64;
-                case TypeCode.Single:
-                    return MarshalType.FP32;
-                case TypeCode.Double:
-                    return MarshalType.FP64;
-                case TypeCode.String:
-                    return MarshalType.STRING;
-                case TypeCode.Char:
-                    return MarshalType.CHAR;
-            }
-
-            if (type.IsArray)
-            {
-                if (!type.IsSZArray)
-                    throw new JSException("Only single-dimensional arrays with a zero lower bound can be marshaled to JS");
-
-                var elementType = type.GetElementType();
-                switch (Type.GetTypeCode(elementType))
-                {
-                    case TypeCode.Byte:
-                        return MarshalType.ARRAY_UBYTE;
-                    case TypeCode.SByte:
-                        return MarshalType.ARRAY_BYTE;
-                    case TypeCode.Int16:
-                        return MarshalType.ARRAY_SHORT;
-                    case TypeCode.UInt16:
-                        return MarshalType.ARRAY_USHORT;
-                    case TypeCode.Int32:
-                        return MarshalType.ARRAY_INT;
-                    case TypeCode.UInt32:
-                        return MarshalType.ARRAY_UINT;
-                    case TypeCode.Single:
-                        return MarshalType.ARRAY_FLOAT;
-                    case TypeCode.Double:
-                        return MarshalType.ARRAY_DOUBLE;
-                    default:
-                        throw new JSException($"Unsupported array element type {elementType}");
-                }
-            }
-            else if (type == typeof(IntPtr))
-                return MarshalType.POINTER;
-            else if (type == typeof(UIntPtr))
-                return MarshalType.POINTER;
-            else if (type == typeof(SafeHandle))
-                return MarshalType.SAFEHANDLE;
-            else if (typeof(Delegate).IsAssignableFrom(type))
-                return MarshalType.DELEGATE;
-            else if ((type == typeof(Task)) || typeof(Task).IsAssignableFrom(type))
-                return MarshalType.TASK;
-            else if (type.FullName == "System.Uri")
-                return MarshalType.URI;
-            else if (type.IsPointer)
-                return MarshalType.POINTER;
-
-            if (type.IsValueType)
-                return MarshalType.VT;
-            else
-                return MarshalType.OBJECT;
-        }
-
-        public static char GetCallSignatureCharacterForMarshalType(MarshalType t, char? defaultValue)
-        {
-            switch (t)
-            {
-                case MarshalType.BOOL:
-                    return 'b';
-                case MarshalType.UINT32:
-                case MarshalType.POINTER:
-                    return 'I';
-                case MarshalType.INT:
-                    return 'i';
-                case MarshalType.UINT64:
-                    return 'L';
-                case MarshalType.INT64:
-                    return 'l';
-                case MarshalType.FP32:
-                    return 'f';
-                case MarshalType.FP64:
-                    return 'd';
-                case MarshalType.STRING:
-                    return 's';
-                case MarshalType.URI:
-                    return 'u';
-                case MarshalType.SAFEHANDLE:
-                    return 'h';
-                case MarshalType.ENUM:
-                    return 'j'; // this is wrong for uint enums
-                case MarshalType.ENUM64:
-                    return 'k'; // this is wrong for ulong enums
-                case MarshalType.TASK:
-                case MarshalType.DELEGATE:
-                case MarshalType.OBJECT:
-                    return 'o';
-                case MarshalType.VT:
-                    return 'a';
-                default:
-                    if (defaultValue.HasValue)
-                        return defaultValue.Value;
-                    else
-                        throw new JSException($"Unsupported marshal type {t}");
-            }
-        }
-
         /// <summary>
         /// Gets the MethodInfo for the Task{T}.Result property getter.
         /// </summary>
@@ -242,15 +93,19 @@ namespace System.Runtime.InteropServices.JavaScript
         /// ensuring that trimming doesn't change the application's behavior.
         /// </remarks>
         [UnconditionalSuppressMessage("ReflectionAnalysis", "IL2070:UnrecognizedReflectionPattern",
-            Justification = "Task<T>.Result is preserved by the ILLinker because _taskGetResultMethodInfo was initialized with it.")]
+            Justification = "Task<T>.Result is preserved by the ILLinker because s_taskGetResultMethodInfo was initialized with it.")]
         public static MethodInfo GetTaskResultMethodInfo(Type taskType)
         {
             if (taskType != null)
             {
-                MethodInfo? result = taskType.GetMethod(TaskGetResultName);
-                if (result != null && result.HasSameMetadataDefinitionAs(s_taskGetResultMethodInfo))
+                if (s_taskGetResultMethodInfo == null)
                 {
-                    return result;
+                    s_taskGetResultMethodInfo = typeof(Task<>).GetMethod(TaskGetResultName);
+                }
+                MethodInfo? getter = taskType.GetMethod(TaskGetResultName);
+                if (getter != null && getter.HasSameMetadataDefinitionAs(s_taskGetResultMethodInfo!))
+                {
+                    return getter;
                 }
             }
 
@@ -269,7 +124,7 @@ namespace System.Runtime.InteropServices.JavaScript
             throw new InvalidProgramException();
         }
 
-        public static async Task<JSObject> ImportAsync(string moduleName, string moduleUrl, CancellationToken cancellationToken )
+        public static async Task<JSObject> ImportAsync(string moduleName, string moduleUrl, CancellationToken cancellationToken)
         {
             Task<JSObject> modulePromise = JavaScriptImports.DynamicImport(moduleName, moduleUrl);
             var wrappedTask = CancelationHelper(modulePromise, cancellationToken);
