+++ src/libraries/System.Runtime.InteropServices.JavaScript/src/System/Runtime/InteropServices/JavaScript/Legacy/DataView.cs
@@ -17,7 +17,7 @@ namespace System.Runtime.InteropServices.JavaScript
         public DataView(ArrayBuffer buffer)
             : base(JavaScriptImports.CreateCSOwnedObject(nameof(DataView), new object[] { buffer }))
         {
-            JSHostImplementation.RegisterCSOwnedObject(this);
+            LegacyHostImplementation.RegisterCSOwnedObject(this);
         }
 
         /// <summary>
@@ -28,7 +28,7 @@ namespace System.Runtime.InteropServices.JavaScript
         public DataView(ArrayBuffer buffer, int byteOffset)
             : base(JavaScriptImports.CreateCSOwnedObject(nameof(DataView), new object[] { buffer, byteOffset }))
         {
-            JSHostImplementation.RegisterCSOwnedObject(this);
+            LegacyHostImplementation.RegisterCSOwnedObject(this);
         }
 
         /// <summary>
@@ -40,7 +40,7 @@ namespace System.Runtime.InteropServices.JavaScript
         public DataView(ArrayBuffer buffer, int byteOffset, int byteLength)
             : base(JavaScriptImports.CreateCSOwnedObject(nameof(DataView), new object[] { buffer, byteOffset, byteLength }))
         {
-            JSHostImplementation.RegisterCSOwnedObject(this);
+            LegacyHostImplementation.RegisterCSOwnedObject(this);
         }
 
         /// <summary>
